Move lock.go functions into iterator.go, lock_refs.go and store.go
This makes the code more structured, and is a precursor to a subsequent commit
of trying to de-duplicate a lot of this logic to reduce API & bug surface
